bach-info-collect/requirements.txt
pandas==2.2.3
openpyxl==3.1.5
deepl==1.15.0
python-dotenv==1.0.0
